Fetch monthly plans from the server

+++ app/src/main/java/com/myplaygroup/app/feature_main/domain/model/MonthlyPlan.kt
@@ -10,5 +10,5 @@ data class MonthlyPlan(
     val paid: Boolean,
     val planName: String,
     val daysOfWeek: List<DayOfWeek>,
-    val planPrice: Int
+    val planPrice: Long
 )
